--- src/main/java/jp/igapyon/simpleodata4/oiyokan/h2/sql/package-info.java
@@ -1,4 +1,0 @@
-/**
- * JDBC用のSQL文を構築するためのクラス群.
- */
-package jp.igapyon.simpleodata4.oiyokan.h2.sql;
