--- src/main/java/jp/igapyon/simpleodata4/package-info.java
@@ -1,4 +1,0 @@
-/**
- * OData サーバを駆動するためのアプリ.
- */
-package jp.igapyon.simpleodata4;
