+++ src/main/java/jp/oiyokan/h2/sql/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * JDBC用のSQL文を構築するためのクラス群.
+ */
+package jp.oiyokan.h2.sql;
