+++ src/main/java/jp/oiyokan/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 要素のデータを構築するクラス群.
+ */
+package jp.oiyokan;
