+++ src/main/java/jp/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * OData サーバを駆動するためのアプリ.
+ */
+package jp;
